% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_methods.R
\name{robs_observe}
\alias{robs_observe}
\title{Add an observer to a \code{robservable} notebook input through \code{robservableProxy}}
\usage{
robs_observe(robs = NULL, observer = NULL)
}
\arguments{
\item{robs}{\code{robservableProxy} that you would like to update}

\item{observer}{\code{character} name(s) of inputs to observe}
}
\value{
\code{robservable_proxy}
}
\description{
Add an observer to a \code{robservable} notebook input through \code{robservableProxy}
}
