% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_methods.R
\name{robs_update}
\alias{robs_update}
\title{Update \code{robservable} through \code{robservableProxy}}
\usage{
robs_update(robs = NULL, ...)
}
\arguments{
\item{robs}{\code{robservableProxy} that you would like to update}

\item{...}{named arguments to represent variables or inputs to update}
}
\value{
\code{robservable_proxy}
}
\description{
Update \code{robservable} through \code{robservableProxy}
}
