% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robservable.R
\name{robservable}
\alias{robservable}
\title{Display an Observable notebook as HTML widget}
\usage{
robservable(
  notebook,
  output_cell = NULL,
  input = NULL,
  input_df = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{notebook}{The notebook id, such as "@d3/bar-chart"}

\item{output_cell}{The name of the cell to be rendered. If NULL,  the whole notebook is rendered}

\item{input}{A named list of cells to be updated.}

\item{input_df}{A named list of d3-array cells to be updated}

\item{width}{htmlwidget width}

\item{height}{htmlwidget height}

\item{elementId}{optional manual widget HTML id}
}
\description{
Display an Observable notebook as HTML widget
}
\details{
Values passed in `input_df` are converted using the JavaScript function `HTMLWidgets.dataframeToD3`.
}
