% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robservable.R
\name{robservable}
\alias{robservable}
\title{Display an Observable notebook as HTML widget}
\usage{
robservable(
  notebook,
  cell = NULL,
  input = NULL,
  observers = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{notebook}{The notebook id, such as "@d3/bar-chart"}

\item{cell}{The name of the cell to be rendered. If NULL,  the whole notebook is rendered}

\item{input}{A named list of cells to be updated.}

\item{observers}{A vector of character strings representing variables in observable that
you would like to set as input values in Shiny.}

\item{width}{htmlwidget width}

\item{height}{htmlwidget height}

\item{elementId}{optional manual widget HTML id}
}
\description{
Display an Observable notebook as HTML widget
}
\details{
Values passed in `input_df` are converted using the JavaScript function `HTMLWidgets.dataframeToD3`.
}
\examples{
\donttest{
## Display a notebook cell
robservable(
  "@d3/bar-chart",
  cell= "chart"
)

## Change cells data with input
robservable(
  "@d3/bar-chart",
  cell= "chart",
  input = list(color = "red", height = 700)
)

## Change data frame cells data
df <- data.frame(table(mtcars$cyl))
names(df) <- c("name", "value")
robservable(
  "@d3/horizontal-bar-chart",
  cell= "chart",
  input = list(data = df)
)
}
}
