% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robservable.R
\name{to_js_date}
\alias{to_js_date}
\title{Convert a Date or POSIXt object to a JS Date format}
\usage{
to_js_date(date)
}
\arguments{
\item{date}{object to be converted}
}
\value{
Numeric value representing the number of milliseconds between Unix Epoch
(1 January 1970 UTC) and \code{date}.
}
\description{
Convert a Date or POSIXt object to a JS Date format
}
